@@ -64,10 +64,10 @@
     "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "^4.0.2",
     "phantomjs-prebuilt": "^2.1.16",
-    "react": "^15.4.0",
+    "react": "^16.12.0",
     "react-addons-css-transition-group": "^15.4.0",
     "react-addons-test-utils": "^15.4.0",
-    "react-dom": "^15.4.0",
+    "react-dom": "^16.12.0",
     "react-immutable-proptypes": "^1.7.0",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
